API documentation

Public API for MID Orthography model

AtticGreek.AtticOrthography — Type

An orthographic system for texts in the pre-Euclidean Attic alphabet.

AtticGreek.atticGreek — Function

Instantiate an AtticOrthography with correct code points and token types.

atticGreek()

PolytonicGreek.sortWords — Function

Alphabetically sort a list of words in Unicode Greek.

sortWords(words, ortho)

Alphabetically sort a list of words in Unicode Greek.

sortWords(words, ortho)

Phonology and syllabification

Exported:

PolytonicGreek.consonants — Function

Identify code points representing consonants.

consonants(ortho)

Implement consonants function of GreekOrthography interface.

consonants(ortho)

Implement GreekOrthography's consonants function for AtticOrthography.

consonants(attic)

PolytonicGreek.vowels — Function

Identify code points representing simple vowels.

vowels(ortho)

This does not include code points representing vowels combined with accents. It does include any code points representing vowels plus breathing or vowels with subscript.

Implement vowels function of GreekOrthography interface.

vowels(ortho)

Implement GreekOrthography's vowels function for AtticOrthography.

vowels(attic)

PolytonicGreek.syllabify — Function

Generic method of GreekOrthography's syllabify function.

syllabify(s, ortho)

Allow syllabification defaulting to LiteraryGreekOrthography.

syllabify(s)

Split string s into an Array of strings representing syllables.

syllabify(s, ortho)

Example

syllables = PolytonicGreek.syllabify("κελεύει")
join(syllables, "-")
"κε-λευ-ει"

Internal:

AtticGreek.alphabetic — Function

Compose a string with all alphabetic characters.

alphabetic()

AtticGreek.punctuation — Function

Compose a string with all punctuation characters.

AtticGreek.isAlphabetic — Function

True if all characters in s are alphabetic.

isAlphabetic(s)

AtticGreek.isPunctuation — Function

True if all characters in s are puncutation.

isPunctuation(s)

AtticGreek.tokenforstring — Function

Create correct OrthographicToken for a given string.

tokenforstring(s)

AtticGreek.splitPunctuation — Function

Split off any trailing punctuation and return an Array of leading string + trailing punctuation.

splitPunctuation(s)

AtticGreek.ultima — Function

Return ultima.

ultima(s, ag)

AtticGreek.penult — Function

Return penult.

penult(s, ag)

AtticGreek.antepenult — Function

Return antepenult.

antepenult(s, ag)

NB: Rms accent since they're not relevant to syllables.

AtticGreek.vowelsonly — Function

Remove all consonants from s.

vowelsonly(s, ag)

```

AtticGreek.longsyllable — Function

True if syll is a long syllable in Attic Greek orthography.

longsyllable(syll, ortho)

Arguments

syll is a single syllable

ortho an instance of AtticOrthography

AtticGreek.finallong — Function

True if syll counts as long for accent in ultima.

finallong(syll, ortho)

AtticGreek.finalshort — Function

True if syll counts as short for accent in ultima.

finalshort(syll, ortho)

Accentuation

Exported:

PolytonicGreek.rmaccents — Function

Remove all accent characters from s.

rmaccents(s, ortho)

Parameters

s is a Greek string

ortho is an implementation of GreekOrthography

Default to using literary Greek orthography for rmaccents.

rmaccents(s)

Implement GreekOrthography's rmaccents function for LiteraryGreekOrthography.

rmaccents(s, ortho)

Implement GreekOrthography's rmaccents function for LiteraryGreekOrthography.

rmaccents(s, ortho)

PolytonicGreek.countaccents — Function

Count accents in s.

countaccents(s, ortho)

Parameters

s is a Greek string

ortho is an implementation of GreekOrthography

Default to using literary Greek orthography for countaccents.

countaccents(s)

Counts number of accents in s.

countaccents(s, ortho)

Counts number of accents in s.

countaccents(s, ortho)

PolytonicGreek.accentword — Function

Accent word s.

accentword(s, placement, ortho)

Parameters

s is a Greek string

placement is :RECESSIVE, or :PENULT

ortho is an implementation of GreekOrthography

Accent word according to specified system of accent placement.

accentword(wrd, placement)
accentword(wrd, placement, ortho)

Parameters

wrd is a string value representing a single lexical token.

placement is one of :RECESSIVE for recessive accent

or :PENULT for persistent accent on the penultimate syllable.

Note that it is not possible to accent the ultima correctly without additional morphological information beyond the string value of the token.

Accent word according to specified system of accent placement.

accentword(wrd, placement, ortho)

Parameters

wrd is a string value representing a single lexical token.

placement is one of :RECESSIVE for recessive accent

or :PENULT for persistent accent on the penultimate syllable.

Note that it is not possible to accent the ultima correctly without additional morphological information beyond the string value of the token.

PolytonicGreek.accentultima — Function

Accent final syllable of wrd.

accentultima(wrd, accent, ortho)

Parameters

s is a Greek string

accent is :ACUTE, or :CIRCUMFLEX

ortho is an implementation of GreekOrthography

Place accent on ultima.

accentultima(wrd, accent)
accentultima(wrd, accent, ortho)

Place specified accent on final syllable of wrd.

accentultima(wrd, accent, ortho)

accent is one of either :ACUTE or :CIRCUMFLEX. The function returns nothing for any other symbol for accent.

PolytonicGreek.accentpenult — Function

Accent penultimate syllable of wrd.

accentpenult(wrd, accent, ortho)

Parameters

s is a Greek string

accent is :ACUTE, or :CIRCUMFLEX

ortho is an implementation of GreekOrthography

Place accent on penult.

accentpenult(wrd, accent)
accentpenult(wrd, accent, ortho)

Place specified accent on penult of wrd.

accentpenult(wrd, accent, ortho)

accent is one of either :ACUTE or :CIRCUMFLEX. The function returns nothing for any other symbol for accent.

PolytonicGreek.accentantepenult — Function

Place default accent(:ACUTE) on antepenult.

accentantepenult(wrd, ortho)

Accent wrd proparoxytone.

accentantepenult(wrd, ortho)

Internal:

AtticGreek.stripenclitic — Function

Remove any second accent due to enclisis from s.

stripenclitic(s, ortho)

AtticGreek.addacute — Function

Add an acute accent to a single vowel or diphthong.

addacute(vowel, ortho)

AtticGreek.addcircumflex — Function

Add a circumflex accent to a single vowel or diphthong

addcircumflex(vowel, ortho)

AtticGreek.accentstripdict — Function

Map accented vowels to unaccented forms, but keep smooth breathing

accentstripdict(ortho)

AtticGreek.flipaccent — Function

Convert grave accent to acute.

flipaccent(s, ag)

AtticGreek.tokenform — Function

Normalize s to a standard form with no enclitics, and all barytone accent converted to oxytone.

tokenform(s, ortho)

AtticGreek.accentsyllable — Function

Add specified accent to a single syllable.

accentsyllable(syll, accent, ortho)

Parameters

syll is a string value for a single syllable.

accent is either :ACUTE or :CIRCUMFLEX. The function returns nothing for any other symbol for accent.