API documentation

Public APIs

AtticGreek.AtticOrthography — Type

An orthographic system for texts in the pre-Euclidean Attic alphabet.

AtticGreek.atticGreek — Function

Instantiate an AtticOrthography with correct code points and token types.

atticGreek()

AtticGreek.tokenizeAtticGreek — Function

Tokenize a string in the Attic Greek alphabet.

tokenizeAtticGreek(s)

PolytonicGreek.consonants — Function

Identify code points representing consonants.

consonants(ortho)

Implement consonants function of GreekOrthography interface.

consonants(ortho)

Implement GreekOrthography's consonants function for AtticOrthography.

consonants(attic)

PolytonicGreek.vowels — Function

Identify code points representing simple vowels.

This does not include code points representing vowels combined with accents. It does include any code points representing vowels plus breathing or vowels with subscript.

Implement vowels function of GreekOrthography interface.

vowels(ortho)

Implement GreekOrthography's vowels function for AtticOrthography.

vowels(attic)

PolytonicGreek.rmaccents — Function

Remove all accent characters from s.

rmaccents(s, ortho)

Parameters

s is a Greek string

ortho is an implementation of GreekOrthography

Default to using literary Greek orthography for rmaccents.

rmaccents(s)

Implement GreekOrthography's rmaccents function for LiteraryGreekOrthography.

rmaccents(s, ortho)

Implement GreekOrthography's rmaccents function for LiteraryGreekOrthography.

rmaccents(s, ortho)

AtticGreek internals

AtticGreek.accentstripdict — Function

Map accented vowels to unaccented forms, but keep smooth breathing

accentstripdict(ortho)

AtticGreek.alphabetic — Function

Compose a string with all alphabetic characters.

alphabetic()

AtticGreek.punctuation — Function

Compose a string with all punctuation characters.

AtticGreek.isAlphabetic — Function

True if all characters in s are alphabetic.

isAlphabetic(s)

AtticGreek.isPunctuation — Function

True if all characters in s are puncutation.

isPunctuation(s)

AtticGreek.tokenforstring — Function

Create correct OrthographicToken for a given string.

tokenforstring(s)

AtticGreek.splitPunctuation — Function

Split off any trailing punctuation and return an Array of leading string + trailing punctuation.

splitPunctuation(s)